Place valence and bevy specific stuff in valence_anvil behind feature flags
